[AOS] refactor: 패키지 이동
