+++ dodamduck_aos/app/src/main/java/org/chosun/dodamduck/ui/navigation/DodamDuckNavigation.kt
@@ -29,7 +29,7 @@ import org.chosun.dodamduck.presentation.trade.TradeScreen
 import org.chosun.dodamduck.presentation.library.LibraryScreen
 import org.chosun.dodamduck.presentation.auth.LoginScreen
 import org.chosun.dodamduck.presentation.auth.OnboardingScreen
-import org.chosun.dodamduck.presentation.post.PostDetailScreen
+import org.chosun.dodamduck.presentation.post.detail.PostDetailScreen
 import org.chosun.dodamduck.presentation.post.PostWriteScreen
 import org.chosun.dodamduck.presentation.auth.RegisterScreen
 import org.chosun.dodamduck.presentation.search.SearchScreen
